Class used by trialr to fit Wason & Seaman's Augmented Binary method.

Class used by trialr to fit Wason & Seaman's Augmented Binary method.

augbin_fit(num_patients, tumour_size, non_shrinkage_failure, fit)

Arguments

num_patients
Integer, the number of patients analysed.

tumour_size
matrix-like object containing tumour size measures, with rows representing patients and columns representing chronological standardised assessment points. Column one is baseline.

non_shrinkage_failure
matrix-like object containing logical indicators of non-shrinkage failure, with rows representing patients and columns representing chronological standardised assessment points.

fit
An object of class stanfit, containing the posterior samples.

References

Wason JMS, Seaman SR. Using continuous data on tumour measurements to improve inference in phase II cancer studies. Statistics in Medicine. 2013;32(26):4639-4650. doi:10.1002/sim.5867

Eisenhauer EA, Therasse P, Bogaerts J, et al. New response evaluation criteria in solid tumours: Revised RECIST guideline (version 1.1). European Journal of Cancer. 2009;45(2):228-247. doi:10.1016/j.ejca.2008.10.026

See also