Calculate codified CRM doses.

Calculate the codified CRM doses that map to probability of toxicity prob_tox in a logistic model with expected values for intercept and gradient. I.e. find \(x[i]\) such that \(logit(p[i]) = \alpha + \beta x[i]\), were \(p\) is prob_tox.

crm_codified_dose_logistic(prob_tox, alpha_mean, beta_mean)

Arguments

prob_tox
Numeric vector, seek codified doses that yield these probabilities of toxicity.

alpha_mean
Numeric, expected value of intercept.

beta_mean
Numeric, expected value of gradient with respect to dose.

Value

Numeric vector of codified doses.

Examples

skeleton <- c(0.05, 0.1, 0.2, 0.5)
crm_codified_dose_logistic(skeleton, 1, 0)

#> [1] -Inf -Inf -Inf -Inf

crm_codified_dose_logistic(skeleton, 3, 0.5)

#> [1] -11.888878 -10.394449  -8.772589  -6.000000