Class of model fit by trialr using the CRM dose-finding design.

Class of model fit by trialr using the CRM dose-finding design.

crm_fit(dose_indices, recommended_dose, prob_tox, median_prob_tox,
  modal_mtd_candidate, prob_mtd, dat, fit)

Arguments

dose_indices
A vector of integers representing the dose-levels under consideration.

recommended_dose
An integer representing the dose-level recommended for the next patient or cohort; or NA if stopping is recommended. The recommended dose typically has associated probability of DLT closest to the target toxicity rate. Contrast to modal_mtd_candidate.

prob_tox
The posterior mean probabilities of toxicity at doses 1:n; a vector of numbers between 0 and 1.

median_prob_tox
The posterior median probabilities of toxicity at doses 1:n; a vector of numbers between 0 and 1.

modal_mtd_candidate
An integer representing the dose-level most likely to be the MTD, i.e. the dose-level that maximises prob_mtd.

prob_mtd
The posterior probability that each dose is the MTD, by the chosen model; a vector of numbers between 0 and 1.

dat
Object crm_params containing data passed to sampling.

fit
An object of class stanfit, containing the posterior samples.

Details

See methods(class = "crm_fit") for an overview of available methods.

See also