Container class for parameters to fit the CRM models in trialr.

Container class for parameters to fit the CRM models in trialr.

crm_params(skeleton, target, a0 = NULL, alpha_mean = NULL,
  alpha_sd = NULL, beta_mean = NULL, beta_sd = NULL, beta_shape = NULL,
  beta_inverse_scale = NULL)

Arguments

skeleton
a vector of the prior guesses of toxicity at doses. This should be a monotonically-increasing vector of numbers between 0 and 1.

target
the target toxicity probability, a number between 0 and 1. This value would normally be one of the values in skeleton, but that is not a requirement.

a0
Value of fixed intercept parameter. Only required for certain models. See Details.

alpha_mean
Prior mean of intercept variable for normal prior. Only required for certain models. See Details.

alpha_sd
Prior standard deviation of intercept variable for normal prior. Only required for certain models. See Details.

beta_mean
Prior mean of gradient variable for normal prior. Only required for certain models. See Details.

beta_sd
Prior standard deviation of slope variable for normal prior. Only required for certain models. See Details.

beta_shape
Prior shape parameter of slope variable for gamma prior. Only required for certain models. See Details.

beta_inverse_scale
Prior inverse scale parameter of slope variable for gamma prior. Only required for certain models. See Details.

Details

Different model parameterisations require that difference parameter values are specified.

Requirements of empiric model

* beta_sd

Requirements of logistic model

* a0 * beta_mean * beta_sd

Requirements of logistic_gamma model

* a0 * beta_shape * beta_inverse_scale

Requirements of logistics model

* a0 * alpha_mean * alpha_sd * beta_mean * beta_sd

See also