Calculate dose-transition pathways for an EffTox study

Calculate dose-transition pathways for an EffTox study. Note that TODO TODO TODO

efftox_dtps_to_dataframe(dat, cohort_sizes, next_dose, ...)

Arguments

dat
An instance of efftox_params, a list of EffTox parameters. An example is yielded by efftox_parameters_demo.

cohort_sizes
vector of future cohort sizes, i.e. positive integers. E.g. To calculate paths for the the next cohort of two followed by another cohort of three, use cohort_sizes = c(2, 3).

next_dose
the dose-level to be given to the immediately next cohort.

...
extra params passed to rstan::sampling.

Value

dose pathways in a data.frame.

References

Brock K, Billingham L, Copland M, Siddique S, Sirovica M, Yap C. Implementing the EffTox dose-finding design in the Matchpoint trial. BMC Medical Research Methodology. 2017;17(1):112. doi:10.1186/s12874-017-0381-x

See also

Examples

# Calculate the paths for the first cohort of 3 in Thall et al 2014 example
dat <- efftox_parameters_demo()
if (FALSE) {
dtps1 <- efftox_dtps_to_dataframe(dat = dat, cohort_sizes = c(3),
                                  next_dose = 1)
}
# To calculate future paths in a partially-observed trial
dat <- efftox_parameters_demo()
dat$doses = array(c(1,1,1))
dat$eff = array(c(0,0,0))
dat$tox = array(c(1,1,1))
dat$num_patients = 3
if (FALSE) {
dtps2 <- efftox_dtps_to_dataframe(dat = dat, cohort_sizes = c(3),
                                  next_dose = 1)
}