Parameters to be passed to EffTox model in Stan

The parameters required by the EffTox Stan model are bundled up in a list, for convenience.

Details

The list object has the following elements:

num_doses, an integer.

real_doses, a vector of numbers.The doses under investigation. They should be ordered from lowest to highest and be in consistent units. E.g., to conduct a dose-finding trial of doses 10mg, 20mg and 50mg, use c(10, 20, 50).

efficacy_hurdle, acceptable doses should have efficacy probability exceeding this value, a number between 0 and 1.

toxicity hurdle, acceptable doses should have toxicity probability less than this value, a number between zero and 1.

p_e, certainty required to infer a dose is acceptable with regards to being probably efficacious; a number between 0 and 1.

p_t, certainty required to infer a dose is acceptable with regards to being probably tolerable; a number between 0 and 1.

p, the p-index of EffTox utility contours. Use efftox_solve_p. A number.

eff0, Efficacy probability required when toxicity is impossible; a number between 0 and 1.

tox1, Toxicity probability permitted when efficacy is guaranteed; a number between 0 and 1.

eff_star, Efficacy probability of a third point on the neutral utility curve that joins (eff0, 0) to (1, tox1). A number between 0 and 1.

tox_star, Toxicity probability of a third point on the neutral utility curve that joins (eff0, 0) to (1, tox1). A number between 0 and 1.

alpha_mean, The prior mean of the intercept term in the toxicity logit model. A number.

alpha_sd, The prior standard deviation of the intercept term in the toxicity logit model. A number.

beta_mean, The prior mean of the slope term in the toxicity logit model. A number.

beta_sd, The prior standard deviation of the slope term in the toxicity logit model. A number.

gamma_mean, The prior mean of the intercept term in the efficacy logit model. A number.

gamma_sd, The prior standard deviation of the intercept term in the efficacy logit model. A number.

zeta_mean, The prior mean of the slope term in the efficacy logit model. A number.

zeta_sd, The prior standard deviation of the slope term in the efficacy logit model. A number.

eta_mean, The prior mean of the squared term coefficient in the efficacy logit model. A number.

eta_sd, The prior standard deviation of the squared term coefficient in the efficacy logit model. A number.

psi_mean, The prior mean of the association term in the combined efficacy-toxicity model. A number.

psi_sd, The prior standard deviation of the association term in the combined efficacy-toxicity model. A number.

eff, A vector of efficacy outcomes for patients 1:num_patients, where 1=efficacy and 0=no efficacy.

tox, A vector of toxicity outcomes for patients 1:num_patients, where 1=toxicity and 0=no toxicity.

doses, A vector of dose-levels given to patients 1:num_patients, where 1=lowest dose, 2=second dose, etc.

num_patients, number of patients, positive integer

An example of the efftox_params list is returned by efftox_parameters_demo.

Value

A list with elements identified above.

See also