Container class for parameters to fit the EffTox model in trialr.

Container class for parameters to fit the EffTox model in trialr.

efftox_params(
  real_doses,
  efficacy_hurdle,
  toxicity_hurdle,
  p_e,
  p_t,
  eff0,
  tox1,
  eff_star,
  tox_star,
  priors
)

Arguments

real_doses
a vector of numbers.The doses under investigation. They should be ordered from lowest to highest and be in consistent units. E.g., to conduct a dose-finding trial of doses 10mg, 20mg and 50mg, use c(10, 20, 50).

efficacy_hurdle
Minimum acceptable efficacy probability. A number between 0 and 1.

toxicity_hurdle
Maximum acceptable toxicity probability. A number between 0 and 1.

p_e
Certainty required to infer a dose is acceptable with regards to being probably efficacious; a number between 0 and 1.

p_t
Certainty required to infer a dose is acceptable with regards to being probably tolerable; a number between 0 and 1.

eff0
Efficacy probability required when toxicity is impossible; a number between 0 and 1 (see Details).

tox1
Toxicity probability permitted when efficacy is guaranteed; a number between 0 and 1 (see Details).

eff_star
Efficacy probability of an equi-utility third point (see Details).

tox_star
Toxicity probability of an equi-utility third point (see Details).

priors
instance of class efftox_priors, the hyperparameters for normal priors on the six model parameters.

See also