Calculate the p-index for EffTox utility contours

Calculate the p-index for EffTox utility contours so that the neutral utility contour intersects the following points in the Prob(Efficacy) - Prob(Toxicity) plane: (eff0, 0), (1, tox1) and (eff_star, tox_star)

efftox_solve_p(eff0, tox1, eff_star, tox_star)

Arguments

eff0
Efficacy probability required when toxicity is impossible; a number between 0 and 1

tox1
Toxicity probability permitted when efficacy is guaranteed; a number between 0 and 1

eff_star
Efficacy probability of an equi-utility third point

tox_star
Toxicity probability of an equi-utility third point

Value

The p-index

References

Thall, Herrick, Nguyen, Venier & Norris. 2014, Effective sample size for computing prior hyperparameters in Bayesian phase I-II dose-finding

Examples

efftox_solve_p(0.5, 0.65, 0.7, 0.25)

#> [1] 0.9773632