Stan model for BEBOP implementation in PePS2 clinical trial

This is the Stan implementation of the BEBOP design for the PePS2 trial, that incorporates baseline predictive information to study co-primary efficacy and toxicity outcomes.

The model is compiled when trialr is installed and is available at run-time under stanmodels$BebopInPeps2.

The design has been submitted for publication by Brock et al. in 2017.

See the BEBOP vignette for a detailed description of the probability model and a demonstration of the pertinent methods implemented in trial.

See also

Examples

# Get model parameters as used in the PePS2 trial.
# This call randomly samples patient outcomes so set a seed
set.seed(123)
dat <- peps2_get_data(num_patients = 60,
                     prob_eff = c(0.167, 0.192, 0.5, 0.091, 0.156, 0.439),
                     prob_tox = rep(0.1, 6),
                     eff_tox_or = rep(1, 6))
# Fit the observed data to the model using rstan
samp <- rstan::sampling(stanmodels$BebopInPeps2, data = dat)

#> trying deprecated constructor; please alert package maintainer

#> 
#> SAMPLING FOR MODEL 'BebopInPeps2' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 0.000326 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 3.26 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 2.46827 seconds (Warm-up)
#>                2.5474 seconds (Sampling)
#>                5.01567 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'BebopInPeps2' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 0.000177 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 1.77 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 2.50689 seconds (Warm-up)
#>                2.41228 seconds (Sampling)
#>                4.91918 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'BebopInPeps2' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 0.000176 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 1.76 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 2.42431 seconds (Warm-up)
#>                2.77913 seconds (Sampling)
#>                5.20345 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'BebopInPeps2' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 0.000176 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 1.76 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 2.50354 seconds (Warm-up)
#>                2.66533 seconds (Sampling)
#>                5.16888 seconds (Total)
#> 

# The fit object is quite crude. Post-process to perform useful inference:
decision <- peps2_process(dat, samp)
decision$Accept   # Accept in cohort 2, 3, 5, 6 but not 1 or 4

#> [1] FALSE  TRUE  TRUE FALSE  TRUE  TRUE

decision$ProbEff  # The probability of efficacy is driving that decision

#> [1] 0.07441455 0.24984030 0.42934468 0.04037307 0.14885626 0.28395153