Parameters to be passed to BebopInPeps2 model in Stan

The parameters required by the BebopInPeps2 Stan model are bundled up in a list, for convenience. This list contains elements that we might call "data" (e.g. trial outcomes), and "parameters" (e.g. modelling choices).

Details

The list object has the following elements:

J, Total number of patients used, positive integer.

eff, a vector of logical values to represent the presence (=1) or absence (=0) of efficacy events in the patients.

tox, a vector of logical values to represent the presence (=1) or absence (=0) of toxicity events in the patients.

x1, a vector of logical values to represent whether the patients are pre-treated (=1) or treatment-naive (=0).

x2, a vector of logical values to represent whether the patients are Low PD-L1 (=1) or not (=0).

x3, a vector of logical values to represent whether the patients are Medium PD-L1 (=1) or not (=0). Logically, if x2 and x3 are both zero, then the patient is High PD-L1.

cohort_n, vector of cohort sizes, provided for convenience.

cohort_eff, vector of counts of efficacy events by cohort, provided for convenience.

cohort_tox, vector of counts of toxicity events by cohort, provided for convenience.

alpha_mean, The prior mean of alpha. Alpha is the efficacy model intercept.

alpha_sd, The prior standard deviation of alpha. Alpha is the efficacy model intercept.

beta_mean, The prior mean of beta. Beta is the efficacy model term for being previously treated.

beta_sd, The prior standard deviation of beta. Beta is the efficacy model term for being previously treated.

gamma_mean, The prior mean of gamma. Gamma is the efficacy model term for being PD-L1 score = Low.

gamma_sd, The prior standard deviation of gamma. Gamma is the efficacy model term for being PD-L1 score = Low.

zeta_mean, The prior mean of zeta. Zeta is the efficacy model term for being PD-L1 score = Medium.

zeta_sd, The prior standard deviation of zeta. Zeta is the efficacy model term for being PD-L1 score = Medium.

lambda_mean, The prior mean of lambda. Lambda is the toxicity model intercept.

lambda_sd, The prior standard deviation of lambda. Lambda is the toxicity model intercept.

psi_mean, The prior mean of psi. Psi is the joint model association parameter.

psi_sd, The prior standard deviation of psi. Psi is the joint model association parameter.

An example of the peps2_params list is returned by peps2_get_data.

Value

A list with elements identified above.

See also