Sample data from the Augmented Binary model prior predictive distribution.

Sample data from the prior predictive distributions of the two-period, single arm Augmented Binary model, subject to chosen prior parameters.

prior_predictive_augbin_2t_1a(
  num_samps,
  alpha_mean,
  alpha_sd,
  beta_mean,
  beta_sd,
  gamma_mean,
  gamma_sd,
  sigma_mean,
  sigma_sd,
  omega_lkj_eta,
  alpha_d1_mean,
  alpha_d1_sd,
  gamma_d1_mean,
  gamma_d1_sd,
  alpha_d2_mean,
  alpha_d2_sd,
  gamma_d2_mean,
  gamma_d2_sd
)

Arguments

num_samps
Number of samples.

alpha_mean
Prior mean of alpha parameter.

alpha_sd
Prior sd of alpha parameter.

beta_mean
Prior mean of beta parameter.

beta_sd
Prior sd of beta parameter.

gamma_mean
Prior mean of gamma parameter.

gamma_sd
Prior sd of gamma parameter.

sigma_mean
Prior mean of sigma parameter.

sigma_sd
Prior sd of sigma parameter.

omega_lkj_eta
Prior eta parameter for LKJ prior on covariance matrix of log tumour sizes.

alpha_d1_mean
Prior mean of alpha_D1 parameter.

alpha_d1_sd
Prior sd of alpha_D1 parameter.

gamma_d1_mean
Prior mean of gamma_D1 parameter.

gamma_d1_sd
Prior sd of gamma_D1 parameter.

alpha_d2_mean
Prior mean of alpha_D2 parameter.

alpha_d2_sd
Prior sd of alpha_D2 parameter.

gamma_d2_mean
Prior mean of gamma_D2 parameter.

gamma_d2_sd
Prior sd of gamma_D2 parameter.

Value

Object of class tibble

See also

Examples

prior_predictive_augbin_2t_1a(num_samps = 1000,
                              alpha_mean = 0, alpha_sd = 1,
                              beta_mean = 0, beta_sd = 1,
                              gamma_mean = 0, gamma_sd = 1,
                              sigma_mean = 0, sigma_sd = 1,
                              omega_lkj_eta = 1,
                              alpha_d1_mean = 0, alpha_d1_sd = 1,
                              gamma_d1_mean = 0, gamma_d1_sd = 1,
                              alpha_d2_mean = 0, alpha_d2_sd = 1,
                              gamma_d2_mean = 0, gamma_d2_sd = 1)

#> # A tibble: 1,000 x 9
#>       id    z0          z1          z2    y0      y1     y2  prob_d1     prob_d2
#>    <int> <dbl>       <dbl>       <dbl> <dbl>   <dbl>  <dbl>    <dbl>       <dbl>
#>  1     1  5.47  10.7         15.2          0   0.672  1.02   1.00e+0     9.86e-1
#>  2     2  9.05   0.334        0.245        0  -3.30  -3.61   9.97e-1     5.01e-1
#>  3     3  9.42 892.        1088.           0   4.55   4.75   2.50e-5     1.00e+0
#>  4     4  6.16   0.00257      0.00437      0  -7.78  -7.25   5.14e-2     3.07e-1
#>  5     5  8.03   0.0508       0.181        0  -5.06  -3.79   1.00e+0     3.39e-1
#>  6     6  5.05   9.74        14.3          0   0.658  1.04   6.91e-2     1.44e-5
#>  7     7  6.84   0.0000369    0.000337     0 -12.1   -9.92   3.64e-2     7.20e-1
#>  8     8  8.49  67.4          3.36         0   2.07  -0.925  1.33e-2     3.68e-8
#>  9     9  9.48   3.96         6.76         0  -0.872 -0.338  1.00e+0     6.65e-1
#> 10    10  8.24   2.09         1.95         0  -1.37  -1.44   9.69e-1     9.70e-1
#> # … with 990 more rows