Sample LKJ correlation matrices.

This function was copied from Richard McElreath's rethinking package hosted at https://github.com/rmcelreath/rethinking. In turn, he appears to have copied it from Ben Bolker's rLKJ function from the emdbook package, although I cannot find it there (else I would have imported it).

rlkjcorr(n, K, eta = 1)

Arguments

n
Number of matrices to sample.

K
dimenstion of matrix to sample.

eta
Distribution parameter

Value

matrix