Sample data from the Augmented Binary model prior predictive distribution.

Sample data from the prior predictive distributions of the two-period, single arm Augmented Binary model, subject to chosen prior parameters.

prior_predictive_augbin_2t_1a(num_samps, alpha_mean, alpha_sd, beta_mean,
  beta_sd, gamma_mean, gamma_sd, sigma_mean, sigma_sd, omega_lkj_eta,
  alpha_d1_mean, alpha_d1_sd, gamma_d1_mean, gamma_d1_sd, alpha_d2_mean,
  alpha_d2_sd, gamma_d2_mean, gamma_d2_sd)

Arguments

num_samps
Number of samples.

alpha_mean
Prior mean of alpha parameter.

alpha_sd
Prior sd of alpha parameter.

beta_mean
Prior mean of beta parameter.

beta_sd
Prior sd of beta parameter.

gamma_mean
Prior mean of gamma parameter.

gamma_sd
Prior sd of gamma parameter.

sigma_mean
Prior mean of sigma parameter.

sigma_sd
Prior sd of sigma parameter.

omega_lkj_eta
Prior eta parameter for LKJ prior on covariance matrix of log tumour sizes.

alpha_d1_mean
Prior mean of alpha_D1 parameter.

alpha_d1_sd
Prior sd of alpha_D1 parameter.

gamma_d1_mean
Prior mean of gamma_D1 parameter.

gamma_d1_sd
Prior sd of gamma_D1 parameter.

alpha_d2_mean
Prior mean of alpha_D2 parameter.

alpha_d2_sd
Prior sd of alpha_D2 parameter.

gamma_d2_mean
Prior mean of gamma_D2 parameter.

gamma_d2_sd
Prior sd of gamma_D2 parameter.

Value

Object of class tibble

See also

Examples

prior_predictive_augbin_2t_1a(num_samps = 1000,
                              alpha_mean = 0, alpha_sd = 1,
                              beta_mean = 0, beta_sd = 1,
                              gamma_mean = 0, gamma_sd = 1,
                              sigma_mean = 0, sigma_sd = 1,
                              omega_lkj_eta = 1,
                              alpha_d1_mean = 0, alpha_d1_sd = 1,
                              gamma_d1_mean = 0, gamma_d1_sd = 1,
                              alpha_d2_mean = 0, alpha_d2_sd = 1,
                              gamma_d2_mean = 0, gamma_d2_sd = 1)

#> # A tibble: 1,000 x 9
#>       id    z0      z1            z2    y0      y1     y2   prob_d1      prob_d2
#>    <int> <dbl>   <dbl>         <dbl> <dbl>   <dbl>  <dbl>     <dbl>        <dbl>
#>  1     1  6.99 1.48e+1    21.6           0   0.751   1.13 0.00441   0.0000000379
#>  2     2  8.94 7.13e-2     0.0183        0  -4.83   -6.19 0.0000203 0.273       
#>  3     3  7.63 2.80e+3  3224.            0   5.90    6.05 0.000104  1           
#>  4     4  5.59 3.55e-3     0.643         0  -7.36   -2.16 0.953     0.530       
#>  5     5  7.06 4.01e+2    21.2           0   4.04    1.10 0.906     1           
#>  6     6  8.33 1.95e+6 42081.            0  12.4     8.53 0.671     1           
#>  7     7  7.26 3.37e-1     0.472         0  -3.07   -2.73 0.0000120 0.245       
#>  8     8  7.93 4.49e-5     0.0000216     0 -12.1   -12.8  0.0000137 0.695       
#>  9     9  8.53 4.99e+1     0.685         0   1.77   -2.52 1.000     1           
#> 10    10  9.16 4.56e-5     0.000575      0 -12.2    -9.68 0.00196   0.438       
#> # … with 990 more rows